requirements.txt
fastapi~=0.96.0
protobuf
protoc-gen-http-python
grpcio-tools
python-dotenv~=1.0.0
SQLAlchemy~=2.0.15
uvicorn~=0.22.0
googleapis-common-protos
googleapis-common-protos-stubs
PyJWT